requirements.txt
torch>=1.9.1
torchvision>=0.3
numpy
opencv-python
matplotlib
kornia>=0.6.11
gradio_imageslider
gradio==4.29.0
opencv-python
laspy==2.3.0
Pillow==9.3.0
spyder-kernels==3.0.3
scikit-learn==1.6.1
alphashape==1.3.1
tqdm==4.64.1
shapely==2.0.1
scipy==1.10.1
numba==0.58.0
trimesh==4.6.4
open3d==0.17.0
jax